+++ lib/src/main/java/com/metaplex/lib/experimental/jen/auctionhouse/Instructions.kt
@@ -6,7 +6,7 @@
 //
 package com.metaplex.lib.experimental.jen.auctionhouse
 
-import com.metaplex.lib.experimental.serialization.format.Borsh
+import com.metaplex.lib.serialization.format.Borsh
 import com.solana.core.AccountMeta
 import com.solana.core.PublicKey
 import com.solana.core.TransactionInstruction
