+++ lib/src/test/java/com/metaplex/lib/serialization/format/BorshSerializationTests.kt
@@ -7,7 +7,6 @@
 
 package com.metaplex.lib.serialization.format
 
-import com.metaplex.lib.experimental.serialization.format.Borsh
 import kotlinx.serialization.*
 import kotlinx.serialization.builtins.serializer
 import org.junit.Assert
